requirements.txt
Django==2.1.5
django-braces==1.13.0
django-crispy-forms==1.7.2
mysqlclient==1.3.14
pytz==2018.9
